+++ src/utils/helper.cpp
@@ -144,7 +144,6 @@ std::string string_printf(const char* format, ...) {
 #endif /* defined (ENABLE_BACKTRACE) */
 
   std::string errmsg("ROCProfiler: fatal error: " + message);
-  fputs(errmsg.c_str(), stderr);
 
   std::cerr << errmsg << std::endl;
   abort();
@@ -251,7 +250,7 @@ void trim(std::string& str) {
 // replace unsuported specail chars with space
 static void handle_special_chars(std::string& str) {
   std::set<char> specialChars = {'!', '@', '#', '$', '%', '&', '(', ')', ',', '*', '+', '-', '.',
-                                 '/', ';', '<', '=', '>', '?', '@', '{', '}', '^', '`', '~', '|'};
+                                 '/', ';', '<', '=', '>', '?', '@', '{', '}', '^', '`', '~', '|', ':'};
 
   // Iterate over the string and replace any special characters with a space.
   for (unsigned int i = 0; i < str.length(); i++) {
@@ -276,6 +275,14 @@ void validate_counters_format(std::vector<std::string>& counters, std::string li
         counters.push_back(counter);
       }
     }
+  }
+
+ // raise exception with correct usage if user still managed to corrupt input
+  for(const auto &itr: counters)
+  {
+    if(!has_counter_format(itr)){
+      rocprofiler::fatal("Bad input metric. usage --> pmc: <counter1> <counter2>");
+    } 
   }
 }